% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{clean_PullAllocations}
\alias{clean_PullAllocations}
\title{clean_PullAllocations}
\usage{
clean_PullAllocations(df, activepage = input$activepage)
}
\arguments{
\item{df}{output of pull allocations}

\item{activepage}{if we are in RAD change things a little from program costing}
}
\description{
This function updates the headers from the main PullAllocations function
}
\examples{
clean_PullAllocations(df,activepage=input$activepage)

}
