% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{update_drillstack_selectInput}
\alias{update_drillstack_selectInput}
\title{update_drillstack_selectInput}
\usage{
update_drillstack_selectInput(session, inputId, choices, selected,
  CostModelInfo = values.setup$CostModelInfo)
}
\arguments{
\item{session}{the session variable used in shiny::updateSelectInput which is called here}

\item{inputId}{id of selectInput element}

\item{choices}{choices for drop down. function will make the Div's speficic to the clients naming conventions}

\item{selected}{What value to select}

\item{CostModelInfo}{Table to use for the naming conventions in the account set-up}
}
\description{
This function updates the selectInputs in the modal of update stack level stacking
}
\examples{
update_drillstack_selectInput(session,inputId,choices,selected,CostModelInfo=values.setup$CostModelInfo)

}
